+++ spring-batch-infrastructure/src/test/java/org/springframework/batch/item/xml/StaxEventItemWriterTests.java
@@ -763,7 +763,7 @@ public class StaxEventItemWriterTests {
 			public void marshal(Object graph, Result result) throws XmlMappingException, IOException {
 				super.marshal(graph, result);
 				try {
-					StaxUtils.getXmlEventWriter(result).close();
+					StaxTestUtils.getXmlEventWriter(result).close();
 				} catch (Exception e) {
 					throw new RuntimeException("Exception while writing to output file", e);
 				}
@@ -946,8 +946,8 @@ public class StaxEventItemWriterTests {
 		public void marshal(Object graph, Result result) throws XmlMappingException, IOException {
 			Assert.isInstanceOf( Result.class, result);
 			try {
-				StaxUtils.getXmlEventWriter( result ).add( XMLEventFactory.newInstance().createStartElement(namespacePrefix, namespace, graph.toString()));
-				StaxUtils.getXmlEventWriter( result ).add( XMLEventFactory.newInstance().createEndElement(namespacePrefix, namespace, graph.toString()));
+				StaxTestUtils.getXmlEventWriter( result ).add( XMLEventFactory.newInstance().createStartElement(namespacePrefix, namespace, graph.toString()));
+				StaxTestUtils.getXmlEventWriter( result ).add( XMLEventFactory.newInstance().createEndElement(namespacePrefix, namespace, graph.toString()));
 			}
 			catch ( Exception e) {
 				throw new RuntimeException("Exception while writing to output file", e);
